requirement.txt
absl-py==2.3.1
accelerate==1.7.0
bert-score==0.3.13
brotlicffi==1.0.9.2
cffi==1.17.1
charset-normalizer==3.3.2
filelock==3.17.0
huggingface-hub==0.34.3
idna==3.7
jinja2==3.1.6
markupsafe==3.0.2
networkx==3.4.2
numpy==1.26.4
packaging==25.0
peft==0.15.2
pillow==11.3.0
protobuf==6.31.1
pycparser==2.21
pysocks==1.7.1
pyyaml==6.0.2
requests==2.32.4
rouge-score==0.1.2
setuptools==78.1.1
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchtriton==3.1.0
torchvision==0.20.1
transformers==4.52.4
typing_extensions==4.12.2
urllib3==2.5.0
wheel==0.45.1
yaml==0.2.5